Move executor code into separate file
